Separate macros and implementation
